+++ packages/python-packages/api-stub-generator/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## Version 0.2.9 (Unreleased)
+Fixed issue where Python 3-style type hints stopped displaying
+  their inner types.
+Fixed issue where docstring annotations were preferred over
+  Python 2-style type hints for return types. Since docstrings
+  have a 2-line limit, this preference didn't make sense.
+
 ## Version 0.2.8 (Unreleased)
 Kwargs that were previously displayed as "type = ..." will now
   be displayed as "Optional[type] = ..." to align with syntax